Refactor cameras app layout
