+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/cameras/api/util_views.py
@@ -8,7 +8,7 @@ from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from .models import Camera
+from ..models import Camera
 
 logger = logging.getLogger(__name__)
